git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@2152 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/app_alogclip/main.cpp
@@ -171,7 +171,7 @@ int main(int argc, char *argv[])
   string lpct = doubleToString(clipped_lines_pct, 2);
   string cpct = doubleToString(clipped_chars_pct, 2);
 
-  string digits = intToString((int)(log10(clipped_chars_total)));
+  string digits = intToString((int)(log10((double)clipped_chars_total)));
 
   printf("\n\n");
